Allow externalizing of Filter configuration.
This allows the filter to be configured via dependency injection (eg Spring's DelegatingFilterProxy) and therefore the externalization of configuration outside of the built application.

Whilst being able to specify a config-location is fine as the code is at the moment, we need to be able to generically configure that location on an environment by environment basis. This is already possible via Spring with regards to hibernate l2 cache and the client definition within the hazelcast namespace, but for the filter we need to be able to specify this outside of the web.xml - this can be accomplished by wiring the Filter as a DelegatingFilterProxy *if* we have a means of specifying the equivalent of the init-parameters outside the FilterConfig - hence the modifications to WebFilter to support a properties file injection.

The Spring definition then becomes eg:

<bean id="filter" class="com.hazelcast.web.WebFilter">
  <constructor-arg>
    <props>
      <prop key="client-config-location">${location.reference}</prop>
      <prop key="use-client">true</prop>
      ...
    </props>
  </constructor-arg>
</bean>

and in web.xml:

<filter>
  <filter-name>filter</filter-name>
  <filter-class>org.springframework.web.filter.DelegatingFilterProxy</filter-class>
  <init-param>
    <param-name>targetFilterLifecycle</param-name>
    <param-value>true</param-value>
  </init-param>
</filter>

This change simply allows the external configuration and doesn't modify any of the existing functionality.