Allow RequestWrapper.RequestDispatcher to use the provided servletRequest
Currently, the RequestDispatcher that's returned by RequestWrapper always passes
the RequestWrapper to the wrapped RequestDispatcher instead of the request that
was provided to the forward or include method. This prevents a caller from
providing a wrapper of its own; the provided wrapper is ignored by the
RequestDispatcher.

The current behaviour is presumably a defensive move to prevent callers from
passing in the unwrapped request and thereby 'escaping' the wrapper that
provides the distributed HttpSession. There is no need for this because:

* if the caller has access to the unwrapped request, he can already break
  the distributed HttpSession
* if the caller does not have access to the unwrapped request, he can't
  break the distributed HttpSession either.