Add automated release workflow files  (#137)
Inspired by
https://github.com/hazelcast/quarkus-hazelcast-client/tree/master/.github/workflows,
this PR adds the workflows needed for semi-automated release workflow:

- A snapshot is created for every push to a master
- After prepare-release is invoked manually with a version, a tag and a
commit that updates the version is pushed by the prepare-release.
- deploy-release is run afterwards because it runs when a tag is pushed.
This workflow will actually deploy the release to remote mvn repository.
This workflow can also be run manually. After this runs, it'll create a
github release as a draft with the release artifacts (sources jar and
actual release jar). Then we'll update the changelog and release it.


The release process of hazelcast-wm will be described in a confluence
page and reviewed by Nish. (wip)

Other changes:
- maven-javadoc-plugin's javaApiLinks properties are removed because the
link was being redirected to java 22 page, and also java 1.6/1.7 is
unsupported now
- Add mvn release plugin used by workflow(s)
- Upgrade mvn source plugin version

---------

Co-authored-by: Oliver Howell <oliver.howell@hazelcast.com>